added layout and separated header, nav and footer sections and added some style
