+++ first-app/app/globals.css
@@ -1,6 +1,7 @@
 :root {
   --background: #ffffff;
   --foreground: #171717;
+  --border-color: #7c7c7c;
 }
 
 @media (prefers-color-scheme: dark) {
