+++ first-app/app/layout.js
@@ -1,16 +1,7 @@
-import { Geist, Geist_Mono } from "next/font/google";
-
 import "./globals.css";
-
-const geistSans = Geist({
-  variable: "--font-geist-sans",
-  subsets: ["latin"],
-});
-
-const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
-});
+import Header from "@/components/header/header";
+import Nav from "@/components/navigation/navigation";
+import Footer from "@/components/footer/footer";
 
 export const metadata = {
   title: "Ali Bahrami's Portfolio",
@@ -39,8 +30,11 @@ export const metadata = {
 export default function RootLayout({ children }) {
   return (
     <html lang="en">
-      <body className={`${geistSans.variable} ${geistMono.variable}`}>        
+      <body className="">
+        <Header /> 
+        <Nav />    
         {children}
+        <Footer />
       </body>
     </html>
   );
