+++ first-app/components/footer/footer.js
@@ -1,0 +1,9 @@
+import styles from './footer.module.css';
+
+export default function Footer() {
+  return (
+    <footer className={styles.footer}>
+        <p>Footer section</p>
+      </footer>
+    );
+}
