+++ first-app/components/footer/footer.module.css
@@ -1,0 +1,9 @@
+.footer {
+    display: flex;
+    justify-content: center;
+    flex-direction: column;
+    align-items: center;
+    gap: 1rem;
+    height: 5rem;
+    border-top: 1px solid var(--border-color);
+}
