+++ first-app/components/header/header.module.css
@@ -4,4 +4,5 @@
     flex-direction: column;
     align-items: center;
     gap: 1rem;
+    height: 8rem;
 }
