+++ first-app/components/navigation/navigation.modules.css
@@ -1,0 +1,7 @@
+.navbar {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 1rem;
+  grid-row-start: 1;
+}
